DynamicSnippets - initial commit
The start of getting the trace output from the LS into a file (OutputTest), to see how it can be used later for the snippets

a quick save of the current progress

Caution: it does not work like this -> still needs some fixes!

some more changes

Splitting of DocumentSymbol

generics, ports, entities and architectures can now be split into their names and ports and generics without a problem! (alt leat with my test entity with generics and ports it worked)